+++ .gitignore
@@ -3,6 +3,7 @@
 # IDE's and text editors #
 # Ruby mine
 /.idea
+/.webstorm
 # Symbol tags (e.g. generated by Atom Editor)
 /tags
 # VSCode configuration
