+++ app/services/licences/fetch-licence.service.js
@@ -5,8 +5,6 @@
  * @module FetchLicenceService
  */
 
-const { ref } = require('objection')
-
 const LicenceModel = require('../../models/licence.model.js')
 
 /**
@@ -32,7 +30,6 @@ async function _data (licence) {
   return {
     ...licence,
     ends: licence.$ends(),
-    licenceHolder: licence.$licenceHolder(),
     licenceName,
     registeredTo
   }
@@ -42,65 +39,20 @@ async function _fetchLicence (id) {
   const result = await LicenceModel.query()
     .findById(id)
     .select([
-      'expiredDate',
-      'id',
       'include_in_presroc_billing',
       'include_in_sroc_billing',
-      'lapsedDate',
       'licenceRef',
+      'expiredDate',
       'revokedDate',
-      'startDate'
+      'lapsedDate'
     ])
-    .withGraphFetched('region')
-    .modifyGraph('region', (builder) => {
-      builder.select([
-        'id',
-        'displayName'
-      ])
-    })
-    .withGraphFetched('permitLicence')
-    .modifyGraph('permitLicence', (builder) => {
-      builder.select([
-        ref('licenceDataValue:data.current_version.purposes').as('purposes')
-      ])
-    })
     .withGraphFetched('licenceDocumentHeader')
     .modifyGraph('licenceDocumentHeader', (builder) => {
       builder.select([
         'licenceDocumentHeaders.id'
       ])
     })
-    .withGraphFetched('licenceVersions')
-    .modifyGraph('licenceVersions', (builder) => {
-      builder.select(['licenceVersions.id'])
-        .where('licenceVersions.status', 'current')
-    })
-    .withGraphFetched('licenceVersions.[licenceVersionPurposes, purposes]')
-    .modifyGraph('[licenceVersionPurposes]', (builder) => {
-      builder.select([
-        'licenceVersionPurposes.abstractionPeriodStartDay',
-        'licenceVersionPurposes.abstractionPeriodStartMonth',
-        'licenceVersionPurposes.abstractionPeriodEndDay',
-        'licenceVersionPurposes.abstractionPeriodEndMonth',
-        'licenceVersionPurposes.licenceVersionId'
-      ])
-    })
-    .modifyGraph('[purposes]', (builder) => {
-      builder.select([
-        'purposes.description'
-      ])
-    })
-    .modify('licenceHolder')
     .modify('registeredToAndLicenceName')
-    .withGraphFetched('licenceGaugingStations')
-    .modifyGraph('licenceGaugingStations', (builder) => {
-      builder.select([
-        'gaugingStations.id',
-        'gaugingStations.label'
-      ])
-        .where('licenceGaugingStations.dateDeleted', null)
-    })
-
   return result
 }
